+++ packages/discovery/src/discovery/output/toDiscoveryOutput.ts
@@ -41,6 +41,7 @@ export function processAnalysis(
           address: x.address,
           unverified: x.isVerified ? undefined : (true as const),
           upgradeability: x.upgradeability,
+          sinceTimestamp: x.deploymentTimestamp.toNumber(),
           values: Object.keys(x.values).length === 0 ? undefined : x.values,
           errors: Object.keys(x.errors).length === 0 ? undefined : x.errors,
           derivedName: x.derivedName,