+++ packages/discovery-types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @l2beat/discovery-types
 
+## 0.3.0
+
+### Minor Changes
+
+- Start tracking the deployment timestamp for contracts
+
 ## 0.2.0
 
 ### Minor Changes